fix: migrate force-revalidate to App Router to eliminate hybrid mode
Moves the Pages Router API route to an App Router Route Handler, removing
the pages/ directory entirely. This fixes the build error:

  Error: <Html> should not be imported outside of pages/_document

which occurs in hybrid mode (App Router + Pages Router) when Next.js
tries to statically generate the Pages Router /404 page.

- Delete: packages/gitbook/src/pages/api/~gitbook/force-revalidate.ts
- Create: packages/gitbook/src/app/~gitbook/force-revalidate/route.ts
- Uses withVerifySignature + revalidatePath (App Router equivalents)

Note: the endpoint URL changes from /api/~gitbook/force-revalidate
to /~gitbook/force-revalidate.
