+++ .changeset/fix-html-hybrid-mode.md
@@ -1,0 +1,5 @@
+---
+"gitbook": patch
+---
+
+Migrate force-revalidate endpoint from Pages Router to App Router to fix hybrid mode `<Html>` build error.
